requirements.txt
gymnasium==0.29.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0.1
tqdm==4.66.1
pandas==2.0.3
colorama==0.4.6
pygame>=2.1.3